+++ libs/stack/core-lib/CHANGELOG.md
@@ -1,3 +1,38 @@
+## [1.13.2](https://github.com/OKAMca/stack/compare/core-lib-v1.13.1...core-lib-v1.13.2) (2024-06-14)
+
+
+### Bug Fixes
+
+* aria labelled by ([6380791](https://github.com/OKAMca/stack/commit/63807918a253462211c35f1119acff64c093027d))
+
+## [1.13.1](https://github.com/OKAMca/stack/compare/core-lib-v1.13.0...core-lib-v1.13.1) (2024-06-11)
+
+
+### Bug Fixes
+
+* publishing for new libs ([#126](https://github.com/OKAMca/stack/issues/126)) ([8750ea6](https://github.com/OKAMca/stack/commit/8750ea643fa22ca06979c8d4030fa2a2d865b86f))
+
+# [1.13.0](https://github.com/OKAMca/stack/compare/core-lib-v1.12.0...core-lib-v1.13.0) (2024-06-11)
+
+
+### Features
+
+* STACK-72-Lib-Create-directus-block ([#115](https://github.com/OKAMca/stack/issues/115)) ([c07db82](https://github.com/OKAMca/stack/commit/c07db82d3b2a483e737e8d0dd07074cfa0b82152))
+
+# [1.12.0](https://github.com/OKAMca/stack/compare/core-lib-v1.11.2...core-lib-v1.12.0) (2024-05-30)
+
+
+### Features
+
+* alerts component ([35ee942](https://github.com/OKAMca/stack/commit/35ee942432edc38d9dd7be9c81992bbe314fba61))
+
+## [1.11.2](https://github.com/OKAMca/stack/compare/core-lib-v1.11.1...core-lib-v1.11.2) (2024-05-29)
+
+
+### Bug Fixes
+
+* select do not work on in nextjs app ([5b04406](https://github.com/OKAMca/stack/commit/5b044064c4f0818255e0b65b8959b2a58f74176f))
+
 ## [1.11.1](https://github.com/OKAMca/stack/compare/core-lib-v1.11.0...core-lib-v1.11.1) (2024-04-04)